+++ neo4j-cypher-dsl/src/main/java/org/neo4j/cypherdsl/core/renderer/DefaultVisitor.java
@@ -18,7 +18,6 @@
  */
 package org.neo4j.cypherdsl.core.renderer;
 
-import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayDeque;
@@ -115,7 +114,7 @@ import org.neo4j.cypherdsl.support.schema_name.SchemaNames;
  */
 @SuppressWarnings({ "unused", "squid:S1172" })
 @RegisterForReflection
-class DefaultVisitor extends ReflectiveVisitor implements RenderingVisitor, Appendable {
+class DefaultVisitor extends ReflectiveVisitor implements RenderingVisitor {
 
 	private static final EnumSet<Operator> SKIP_SPACES = EnumSet.of(Operator.EXPONENTIATION, Operator.UNARY_MINUS, Operator.UNARY_PLUS);
 
@@ -878,9 +877,15 @@ class DefaultVisitor extends ReflectiveVisitor implements RenderingVisitor, Appe
 
 	void enter(Use use) {
 		builder.append("USE ");
+		if (use.dynamic()) {
+			builder.append("graph.byName(");
+		}
 	}
 
 	void leave(Use use) {
+		if (use.dynamic()) {
+			builder.append(")");
+		}
 		builder.append(" ");
 	}
 
@@ -904,20 +909,5 @@ class DefaultVisitor extends ReflectiveVisitor implements RenderingVisitor, Appe
 	protected final String escapeIfNecessary(String potentiallyNonIdentifier) {
 
 		return SchemaNames.sanitize(potentiallyNonIdentifier, false).orElse(null);
-	}
-
-	@Override
-	public Appendable append(CharSequence csq) throws IOException {
-		return builder.append(csq);
-	}
-
-	@Override
-	public Appendable append(CharSequence csq, int start, int end) throws IOException {
-		return builder.append(csq, start, end);
-	}
-
-	@Override
-	public Appendable append(char c) throws IOException {
-		return builder.append(c);
 	}
 }
